+++ lldb/test/API/functionalities/data-formatter/data-formatter-stl/generic/deque/main.cpp
@@ -26,6 +26,11 @@ template <typename T> T fill(T deque) {
   return deque;
 }
 
+void by_ref_and_ptr(std::deque<int> &ref, std::deque<int> *ptr) {
+  puts("stop here");
+  return;
+}
+
 int main() {
   std::deque<int> empty;
   std::deque<int> deque_1 = {1};
@@ -37,5 +42,23 @@ int main() {
   std::deque<Foo_large> deque_200_large;
   deque_200_large = fill<std::deque<Foo_large>>(deque_200_large);
 
-  return empty.size() + deque_1.front() + deque_3.front(); // break here
+  puts("break here");
+
+  empty.push_back(1);
+  puts("break here");
+
+  (empty.push_back(12));
+  (empty.push_back(123));
+  (empty.push_back(1234));
+  (empty.push_back(12345));
+  (empty.push_back(123456));
+  (empty.push_back(1234567));
+  puts("break here");
+
+  by_ref_and_ptr(empty, &empty);
+
+  empty.clear();
+  puts("break here");
+
+  return empty.size() + deque_1.front() + deque_3.front();
 }
