--- lldb/test/API/functionalities/data-formatter/data-formatter-stl/libcxx/deque/Makefile
@@ -1,4 +1,0 @@
-USE_LIBCPP := 1
-CXX_SOURCES := main.cpp
-
-include Makefile.rules
